+++ src/Gate.IO.Api/EarnUni/Enums/GateEarnUniInterestRecordStatus.cs
@@ -1,0 +1,17 @@
+namespace Gate.IO.Api.EarnUni;
+
+/// <summary>
+/// EarnUni interest record status
+/// </summary>
+public enum GateEarnUniInterestRecordStatus : byte
+{
+    /// <summary>
+    /// Failed
+    /// </summary>
+    Failed = 0,
+
+    /// <summary>
+    /// Success
+    /// </summary>
+    Success = 1,
+}